pyproject.toml
[tool.poetry]
name = "hierarchical_data"
version = "0.1.0"
description = "Dropbox-esque Web Interface using Django to create files and folders"
authors = ["Cheria Artis"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
